requirements.txt
annotated-types==0.7.0
blis==1.0.1
boto3==1.35.33
botocore==1.35.33
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
confection==0.1.5
cymem==2.0.8
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.0.2
packaging==24.1
preshed==3.0.9
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
s3transfer==0.10.2
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
textblob==0.18.0.post0
thinc==8.3.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
vaderSentiment==3.3.2
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
